Shim generator for exporting specific Cygwin tools
Cygwin, like any other Unix system, installs executables to global
binary directories (/usr/bin). It is useful to be able to expose
specific binaries from bin to the Command Processor and PowerShell.

GCC in particular expects to be invoked from where it was configured to
be installed, since it uses relative paths from there to find the
auxiliary programs it drives.

This project provides a tiny C launcher which can be used to create shim
executables for invoking Cygwin programs. It sets up the environment to
allow them to be called without realising that they are not running from
a Cygwin shell.

These shims can be put in a different directory which can be added to
PATH, allowing them to be used without pulling in the whole of Cygwin's
bin directory.